requirements.txt
# File location: requirements.txt

snowflake-connector-python
python-dotenv
